feature(ui-ux): rename mock api and add absolute path
